Add tokenization of all primitive type names
This takes the complete list (including its ordering) from the Avro
compiler source code, available at:

https://github.com/apache/avro/blob/master/lang/java/compiler/src/main/javacc/org/apache/avro/compiler/idl/idl.jj

It also includes 'void' as a primitive type, though it's only actually
allowed as a synonym for 'null' when it's a message return type. It
makes more sense to accept this in the parser and then reject it later
during semantic analysis, where we can offer a fix-it and an explanation
for why it's invalid.